portal/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-allauth==65.9.0
django-oauth-toolkit==3.0.1
django-types==0.20.0
djangorestframework==3.16.0
drf-nested-routers==0.94.2
idna==3.10
jwcrypto==1.5.6
jwt==1.3.1
oauthlib==3.2.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
types-psycopg2==2.9.21.20250516
typing_extensions==4.14.0
urllib3==2.4.0
